Custom music sorting bugfix
Custom music sorting used the name "category" instead of "kind" when building the custom music lists.
My fix changes this so that music sorting when separating tracks works correctly again. No need to update/change the config.txt or anything else either.

Also made a small change so that you don't get an error if trying to separate cutscene/title music from the rest when you have 0 of either.
The fix is kinda silly and could probably be better, but it's simple enough and works for now.